requirements.txt
websocket-client~=1.2.3
influxdb-client~=1.24.0
python-dotenv~=0.19.2